Move r.viewadapters package to r.views package
